+++ CHANGES.md
@@ -2,6 +2,7 @@ master (in development)
 =======================
 
 - Fix warning about parsing messages [GH-1]
+- Fix error column offsets [GH-2]
 
 0.1 (Nov 30, 2014)
 ==================